Implement BLS12-381 signature scheme
Implementation is based on Java implementation of Milagro library
https://github.com/milagro-crypto/milagro-crypto-java

However, for key generation Bouncy Castle is used which seems
pretty reliable for the moment. There is an opportunity to replace
all curve math with Bouncy Castle implementation except pairing.
This is the only thing that is not yet in BC library.